refactor(data-explorer): satisfy quality budgets
