+++ package.json
@@ -38,7 +38,7 @@
     "test:chat-thread": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts chat-thread && bun test web/components/chat/chat-thread.dom.test.ts",
     "test:command-failure": "bun test web/components/shared/command-failure.test.ts web/components/shared/command.test.ts",
     "test:code-block": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts code-block && bun test web/components/shared/code-block.dom.test.ts",
-    "test:data-explorer": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts data-explorer && bun test --parallel=2 web/components/data/data-explorer.dom.test.ts web/components/data/data-explorer-retention.dom.test.ts web/components/data/data-explorer-url.test.ts web/components/data/data-explorer-controller.test.ts web/components/data/data-explorer-spec.test.ts",
+    "test:data-explorer": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts data-explorer && bun test --parallel=1 web/components/data/data-explorer-url.test.ts web/components/data/data-explorer-controller.test.ts web/components/data/data-explorer-spec.test.ts && bun test --parallel=1 web/components/data/data-explorer.dom.test.ts && bun test --parallel=1 web/components/data/data-explorer-visualization.dom.test.ts && bun test --parallel=1 web/components/data/data-explorer-retention.dom.test.ts",
     "test:project-page": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts project-page && bun test web/components/project/project-page.dom.test.ts",
     "test:windowed-table": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts windowed-table && bun test web/components/shared/windowed-table.dom.test.ts",
     "test:filter-menu": "node scripts/ensure_playwright.mjs && bun scripts/build_test_assets.ts filter-menu && bun test web/components/shared/filter-menu.dom.test.ts",
